Updated Script
Fixed Several issues.  now it puts the Country of origin in the series folder. which can also be manually added to the TVNetwork.csv file. which was also updated. also fixed Specials so they go to their own folder. always in the finished folder. removed the regex check for keywords in series and folder name. left it to the file name. may add a few back if neccessary over time. not sure what else i did. a few other things though.